Continuing tidying up of the code, I will try to remove the orphan functions soon.
I have confirmed vs the measured data that both the time domain and frequency domain generate results consistent with the measurements when normalised. I will include these as `tests` in the next commit as well as examples of how to use the scattering function in the time and frequency domain.